moved Log/ to CommonLogic/Log/
