+++ CommonLogic/Log/Handlers/DebugMessageFileHandler.php
@@ -8,43 +8,36 @@ use exface\Core\Interfaces\iCanGenerateDebugWidgets;
 use exface\Core\Interfaces\LogHandlerInterface;
 use exface\Core\Widgets\DebugMessage;
 use FemtoPixel\Monolog\Handler\CsvHandler;
-use Monolog\Logger;
 
 class DebugMessageFileHandler implements LogHandlerInterface {
 	private $dir;
-	private $minLogLevel;
-	private $staticFilenamePart;
 
 	/**
 	 * DebugMessageFileHandler constructor.
 	 *
 	 * @param $dir
-	 * @param $staticFilenamePart
-	 * @param $minLogLevel
 	 */
-	function __construct($dir, $staticFilenamePart, $minLogLevel = Logger::DEBUG) {
-		$this->dir                = $dir;
-		$this->staticFilenamePart = $staticFilenamePart;
-		$this->minLogLevel        = $minLogLevel;
+	function __construct( $dir ) {
+		$this->dir = $dir;
 	}
 
-	public function handle($level, $message, array $context = array(), iCanGenerateDebugWidgets $sender = null) {
-		$fileName = $context["id"] . $this->staticFilenamePart;
-		if (!$fileName) {
+	public function handle( $level, $message, array $context = array(), iCanGenerateDebugWidgets $sender = null ) {
+		$fileName = $context["id"];
+		if ( ! $fileName ) {
 			return;
 		}
 
-		$logger = new \Monolog\Logger("Stacktrace");
-		$logger->pushHandler(new CsvHandler($this->dir . "/" . $fileName, $this->minLogLevel));
+		$logger = new \Monolog\Logger( "Stacktrace" );
+		$logger->pushHandler( new CsvHandler( $this->dir . "/" . $fileName, $level ) );
 
-		$debugWidget = $sender->create_debug_widget($this->createDebugMessage());
-		$logger->log($level, $debugWidget->export_uxon_object()->to_json(), $this->prepareContext($context));
+		$debugWidget = $sender->create_debug_widget( $this->createDebugMessage() );
+		$logger->error( $debugWidget->export_uxon_object()->to_json(), $this->prepareContext( $context ) );
 	}
 
-	protected function prepareContext($context) {
+	protected function prepareContext( $context ) {
 		// do not log the exception in this handler
-		if (isset($context["exception"])) {
-			unset($context["exception"]);
+		if ( isset( $context["exception"] ) ) {
+			unset( $context["exception"] );
 		}
 
 		return $context;
@@ -53,10 +46,10 @@ class DebugMessageFileHandler implements LogHandlerInterface {
 	protected function createDebugMessage() {
 		global $exface;
 		$ui   = $exface->ui();
-		$page = UiPageFactory::create_empty($ui);
+		$page = UiPageFactory::create_empty( $ui );
 
-		$debugMessage = new DebugMessage($page);
-		$debugMessage->set_meta_object($page->get_workbench()->model()->get_object('exface.Core.ERROR'));
+		$debugMessage = new DebugMessage( $page );
+		$debugMessage->set_meta_object( $page->get_workbench()->model()->get_object( 'exface.Core.ERROR' ) );
 
 		return $debugMessage;
 	}
