+++ CommonLogic/Log/Handlers/LogfileHandler.php
@@ -3,11 +3,9 @@
 namespace exface\Core\CommonLogic\Log\Handlers;
 
 
-use exface\Core\CommonLogic\Log\Handlers\monolog\AbstractMonologHandler;
 use exface\Core\CommonLogic\Log\Processors\IdProcessor;
 use FemtoPixel\Monolog\Handler\CsvHandler;
 use Monolog\Logger;
-use exface\Core\Interfaces\LogHandlerInterface;
 
 class LogfileHandler extends AbstractMonologHandler {
 	/**
@@ -20,7 +18,7 @@ class LogfileHandler extends AbstractMonologHandler {
 	 * @throws \Exception                If a missing directory is not buildable
 	 * @throws \InvalidArgumentException If stream is not a resource or string
 	 */
-	function __construct($name, $stream, $level = LogHandlerInterface::DEBUG, $bubble = true, $filePermission = null, $useLocking = false) {
+	function __construct($name, $stream, $level = Logger::DEBUG, $bubble = true, $filePermission = null, $useLocking = false) {
 		$logger = new Logger($name);
 		$logger->pushHandler(new CsvHandler($stream, $level, $bubble, $filePermission, $useLocking));
 		$logger->pushProcessor(new IdProcessor());
