+++ CommonLogic/Log/Log.php
@@ -4,9 +4,6 @@ namespace exface\Core\CommonLogic\Log;
 
 
 use exface\Core\CommonLogic\Filemanager;
-use exface\Core\CommonLogic\Log\Handlers\DebugMessageFileHandler;
-use exface\Core\CommonLogic\Log\Handlers\limit\FileLimitingLogHandler;
-use exface\Core\CommonLogic\Log\Handlers\limit\DirLimitingLogHandler;
 use exface\Core\CommonLogic\Log\Handlers\LogfileHandler;
 use exface\Core\Interfaces\LoggerInterface;
 
@@ -22,42 +19,31 @@ class Log
     {
         $logger = new Logger();
 
-        $coreLogFilePath = static::getCoreLogPath();
-	    $detailsLogBasePath = static::getDetailsLogPath();
+        $coreLogPath = static::getCoreLogPath();
+	    $detailsLogPath = static::getDetailsLogPath();
 
-	    $minLogLevel = static::getWorkbench()->get_config()->get_option('LOG.MINIMUM_LEVEL_TO_LOG');
-	    $maxDaysToKeep = static::getWorkbench()->get_config()->get_option('LOG.MAX_DAYS_TO_KEEP');
-	    $detailsStaticFilenamePart = static::getWorkbench()->get_config()->get_option('LOGDETAILS.STATIC_FILENAME_PART');
-
-        $logger->pushHandler(new FileLimitingLogHandler(
-        	function($filename) use ($minLogLevel) {
-	            return new LogfileHandler("exface", $filename, $minLogLevel);
-			}, $coreLogFilePath, $maxDaysToKeep
-        ));
-
+        $logger->pushHandler(new LogfileHandler("exface", $coreLogPath));
         // TODO tvw enable when log details are used
-        $logger->pushHandler(new DirLimitingLogHandler(
-        	function() use ($detailsLogBasePath, $detailsStaticFilenamePart, $minLogLevel) {
-        		return new DebugMessageFileHandler($detailsLogBasePath . DIRECTORY_SEPARATOR . 'details', $detailsStaticFilenamePart, $minLogLevel);
-            }, $detailsLogBasePath, $detailsStaticFilenamePart, $maxDaysToKeep
-        ));
+//        $logger->pushHandler(new DebugMessageFileHandler($detailsLogPath));
 
         return $logger;
     }
 
 	private static function getCoreLogPath() {
-		$workbench = static::getWorkbench();
+		global $exface;
+		$workbench = $exface->ui()->get_workbench();
 
 		$basePath = Filemanager::path_normalize($workbench->filemanager()->get_path_to_base_folder());
 
 		$obj = $workbench->model()->get_object('exface.Core.LOG');
 		$relativePath = $obj->get_data_address();
 
-		return $basePath . '/' . $relativePath . '/log';
+		return $basePath . '/' . $relativePath;
 	}
 
 	private static function getDetailsLogPath() {
-		$workbench = static::getWorkbench();
+		global $exface;
+		$workbench = $exface->ui()->get_workbench();
 
 		$basePath = Filemanager::path_normalize($workbench->filemanager()->get_path_to_base_folder());
 
@@ -65,10 +51,5 @@ class Log
 		$relativePath = $obj->get_data_address();
 
 		return $basePath . '/' . $relativePath;
-	}
-
-	private static function getWorkbench() {
-		global $exface;
-		return $exface->ui()->get_workbench();
 	}
 }
