+++ CommonLogic/Log/Logger.php
@@ -2,11 +2,12 @@
 
 namespace exface\Core\CommonLogic\Log;
 
+
+use exface\Core\Exceptions\LogicException;
 use exface\Core\Interfaces\iCanGenerateDebugWidgets;
 use exface\Core\Interfaces\LoggerInterface;
 use exface\Core\Interfaces\LogHandlerInterface;
 use exface\Core\CommonLogic\Log\Handlers\LogfileHandler;
-use exface\Core\Exceptions\UnderflowException;
 
 class Logger implements LoggerInterface
 {
@@ -171,13 +172,12 @@ class Logger implements LoggerInterface
     /**
      * Pops a handler from the top of the stack and returns it
      *
-     * @throws UnderflowException if no handlers registered
      * @return LogHandlerInterface
      */
     public function popHandler()
     {
         if (!$this->handlers) {
-            throw new UnderflowException('Can not pop handler from an empty handler stack.');
+            throw new LogicException('Can not pop handler from an empty handler stack.');
         }
 
         return array_shift($this->handlers);
