--- Log/Processors/IdProcessor.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace exface\Core\Log\Processors;
-
-
-class IdProcessor
-{
-	/**
-	 * @param  array $record
-	 * @return array
-	 */
-	public function __invoke(array $record)
-	{
-		$idArray = array('id' => $record['context']['id']);
-		unset($record['context']['id']);
-
-		return $idArray + $record;
-	}
-}
